fixing logic to check if node is already registered

+++ manifests/config.pp
@@ -14,35 +14,38 @@ class gcsv5::config() {
         owner  => root,
     }
 
-    exec {'gcsv5_deployment_key':
-        command  =>  "echo \'${gcsv5::deployment_key}\' > ${gcsv5::globus_conf}/deployment-key.json",
-        cwd      =>  "$gcsv5::globus_conf",
-        path     =>  '/bin:/usr/bin:/sbin:/usr/sbin',
-        provider =>  "shell",
-    }
-
-    if ( ! empty($gcsv5::node_info) ) {
-        exec {'gcsv5_node_info':
-            command  =>  "echo \'${gcsv5::node_info}\' > ${gcsv5::globus_conf}/node_info.json",
+    $ginfofile = find_file($gcsv5::globus_info_file)
+    if (! $ginfofile) {
+        exec {'gcsv5_deployment_key':
+            command  =>  "echo \'${gcsv5::deployment_key}\' > ${gcsv5::globus_conf}/deployment-key.json",
             cwd      =>  "$gcsv5::globus_conf",
             path     =>  '/bin:/usr/bin:/sbin:/usr/sbin',
             provider =>  "shell",
         }
-
-        exec {'gcsv5_node_setup_import':
-            command  => "LC_ALL=en_US.utf8 ${gcsv5::gcs_cmd} node setup --client-id ${gcsv5::client_id} --secret ${gcsv5::client_secret} --import-node ./node_info.json --ip-address ${gcsv5::ip_addr} && rm -f ./deployment-key.json && rm -f ./node_info.json",
-            cwd      => '/root/globus_conf',
-            path     => '/bin:/usr/bin:/sbin:/usr/sbin',
-            unless   => "if [[ `/usr/bin/ps -eaf|/usr/bin/grep gridftp|/usr/bin/grep -v grep |/usr/bin/wc -l` -gt 0 ]]; then exit 0; else exit 1;fi;",
-            provider => 'shell'
-        }
-    } else {
-        exec {'gcsv5_node_setup':
-            command  => "${gcsv5::gcs_cmd} node setup --client-id ${gcsv5::client_id} --secret ${gcsv5::client_secret} --ip-address ${gcsv5::ip_addr} --export-node ./node_info_new.json && rm -rf ./deployment-key.json",
-            cwd      => '/root/globus_conf',
-            path     => '/bin:/usr/bin:/sbin:/usr/sbin',
-            unless   => "if [[ `/usr/bin/ps -eaf|/usr/bin/grep gridftp|/usr/bin/grep -v grep |/usr/bin/wc -l` -gt 0 ]]; then exit 0; else exit 1;fi;",
-            provider => 'shell'
+    
+        if ( ! empty($gcsv5::node_info) ) {
+            exec {'gcsv5_node_info':
+                command  =>  "echo \'${gcsv5::node_info}\' > ${gcsv5::globus_conf}/node_info.json",
+                cwd      =>  "$gcsv5::globus_conf",
+                path     =>  '/bin:/usr/bin:/sbin:/usr/sbin',
+                provider =>  "shell",
+            }
+    
+            exec {'gcsv5_node_setup_import':
+                command  => "LC_ALL=en_US.utf8 ${gcsv5::gcs_cmd} node setup --client-id ${gcsv5::client_id} --secret ${gcsv5::client_secret} --import-node ./node_info.json --ip-address ${gcsv5::ip_addr} && rm -f ./deployment-key.json && rm -f ./node_info.json",
+                cwd      => '/root/globus_conf',
+                path     => '/bin:/usr/bin:/sbin:/usr/sbin',
+                unless   => "if [[ `/usr/bin/ps -eaf|/usr/bin/grep gridftp|/usr/bin/grep -v grep |/usr/bin/wc -l` -gt 0 ]]; then exit 0; else exit 1;fi;",
+                provider => 'shell'
+            }
+        } else {
+            exec {'gcsv5_node_setup':
+                command  => "${gcsv5::gcs_cmd} node setup --client-id ${gcsv5::client_id} --secret ${gcsv5::client_secret} --ip-address ${gcsv5::ip_addr} --export-node ./node_info_new.json && rm -rf ./deployment-key.json",
+                cwd      => '/root/globus_conf',
+                path     => '/bin:/usr/bin:/sbin:/usr/sbin',
+                unless   => "if [[ `/usr/bin/ps -eaf|/usr/bin/grep gridftp|/usr/bin/grep -v grep |/usr/bin/wc -l` -gt 0 ]]; then exit 0; else exit 1;fi;",
+                provider => 'shell'
+            }
         }
     }
